+++ src/main/java/com/majorbit/bozza_proj_turni_prenotazioni/application/dto/PostoDTO.java
@@ -17,6 +17,6 @@ public class PostoDTO implements Serializable {
 
     private String nome;
     private boolean disponibile;
-    private Long stanza;
+    private Integer stanza;
 
 }
