+++ src/main/java/com/majorbit/bozza_proj_turni_prenotazioni/domain/repository/PrenotazioneRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import java.sql.Date;
 import java.util.List;
 
-public interface PrenotazioneRepository extends JpaRepository<Prenotazione, Long> {
+public interface PrenotazioneRepository extends JpaRepository<Prenotazione, Integer> {
     @Query("SELECT p FROM Prenotazione p " + "WHERE p.posto = :posto "
             + "AND p.dataInizio BETWEEN :dataInizio AND :dataFine "
             + "AND p.dataFine BETWEEN :dataInizio AND :dataFine")
