+++ src/main/java/com/majorbit/bozza_proj_turni_prenotazioni/domain/repository/StanzaRepository.java
@@ -3,5 +3,5 @@ package com.majorbit.bozza_proj_turni_prenotazioni.domain.repository;
 import com.majorbit.bozza_proj_turni_prenotazioni.domain.model.Stanza;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StanzaRepository extends JpaRepository<Stanza, Long> {
+public interface StanzaRepository extends JpaRepository<Stanza, Integer> {
 }
