+++ src/main/java/com/majorbit/bozza_proj_turni_prenotazioni/domain/repository/UtenteRepository.java
@@ -3,5 +3,5 @@ package com.majorbit.bozza_proj_turni_prenotazioni.domain.repository;
 import com.majorbit.bozza_proj_turni_prenotazioni.domain.model.Utente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UtenteRepository extends JpaRepository<Utente, Long> {
+public interface UtenteRepository extends JpaRepository<Utente, Integer> {
 }
