+++ src/main/java/com/majorbit/bozza_proj_turni_prenotazioni/infrastructure/repository/JPAPianoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JPAPianoRepository extends JpaRepository<Piano, Long>, PianoRepository {
+public interface JPAPianoRepository extends JpaRepository<Piano, Integer>, PianoRepository {
 }
