+++ src/main/java/com/majorbit/bozza_proj_turni_prenotazioni/infrastructure/repository/JPASedeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JPASedeRepository extends JpaRepository<Sede, Long>, SedeRepository {
+public interface JPASedeRepository extends JpaRepository<Sede, Integer>, SedeRepository {
 
 
 }
